CLEANUP: Update Patchwork module for relocation of theory library.
